requirements.txt
finnhub-python
google-adk
mcp
python-dotenv
termcolor
google-generativeai
